dashboard/requirements.txt
streamlit
pandas
numpy
matplotlib
psycopg2-binary
python-dotenv
